refactored code to set state in newslist. delete functionality working
