+++ src/components/ApplicationViews.js
@@ -10,32 +10,12 @@ export default class ApplicationViews extends Component {
 
   isAuthenticated = () => (sessionStorage.getItem("credentials") !== null || localStorage.getItem("credentials") !== null)
 
-  state = {
-    users: [],
-    events: [],
-    todos: [],
-    news: [],
-    messages: [],
-    friends: []
-  }
+  getAllUsers = () => APIManager.getAllEntries("users")
 
-  componentDidMount() {
-    const newState = {}
 
-    APIManager.getAllEntries("users")
-      .then(users => newState.users = users)
-      .then(() => APIManager.getAllEntries("events"))
-      .then(events => newState.events = events)
-      .then(() => APIManager.getAllEntries("todos"))
-      .then(todos => newState.todos = todos)
-      .then(() => APIManager.getAllEntries("news"))
-      .then(news => newState.news = news)
-      .then(() => APIManager.getAllEntries("messages"))
-      .then(messages => newState.messages = messages)
-      .then(() => APIManager.getAllEntries("friends"))
-      .then(friends => newState.friends = friends)
-      .then(() => this.setState(newState))
-  }
+
+
+
 
 
 render() {
@@ -43,7 +23,7 @@ render() {
     <React.Fragment>
       <Route exact path="/news" render={(props) => {
         if (this.isAuthenticated()) {
-          return <NewsList news={this.state.news} deleteEntry={this.deleteEntry} />
+          return <NewsList getAllUsers={this.getAllUsers} />
         } else {
           return <Redirect to="/login" />
         }
